feat!: move database list from xdg cache to state
The connection configuration is part of the configuration, it shouldn't
be wiped along with the cache.
One should be able to  wipe ~/.cache  without annoyance but this makes
it harder.

There is now the problem of migrating the file that should be conveyed
to the user, maybe via a "breaking changes" issue.